Add MIT License
This commit adds the MIT License to the project. The license grants permission to anyone obtaining a copy of the software to use, modify, merge, publish, distribute, sublicense, and/or sell copies of the software, with the condition that the copyright notice and permission notice are included in all copies or substantial portions of the software. The software is provided "as is" without any warranty.